Refactor 'fetch' subcommand code.
