+++ main.py
@@ -10,10 +10,11 @@ import re
 
 import requests
 
+from fetch.point_transfer import transfer_points
 from util.definitions import GITLAB_ISSUE_URL_REGEX
-from util.gitlab_id import get_project_id
 from util.gitlab_issue import get_issues_from_epics, get_issues_from_iterations, \
       get_issues_from_milestones, get_issues_from_projects, get_issue_info
+from util.thunderdome_plan import get_plans
 
 
 def parse_args() -> argparse.Namespace:
@@ -94,7 +95,7 @@ def main() -> None:
     logging.basicConfig(level=logging.INFO)
 
     if args.command == "fetch":
-        plans = fetch_plans(args.battleid, args.api_key)
+        plans = get_plans(args.battleid, args.api_key)
         transfer_points(plans, args.token, args.overwrite)
 
     elif args.command == "create":
@@ -106,151 +107,6 @@ def main() -> None:
 
         create_game(plans, args)
 
-
-def fetch_plans(battle_id: str, api_key: str) -> list[dict]:
-    """
-    Get plans from the Thunderdome API.
-
-    :param battle_id: Battle ID to fetch.
-    :param api_key: API key for the Thunderdome API.
-    :return: Plans for the battle.
-    """
-    logging.info("Fetching plans for battle %s...", battle_id)
-
-    thunderdome_headers = {
-        "accept": "application/json",
-        "X-API-Key": api_key,
-    }
-
-    # Thunderdome request
-    thunderdome_response = requests.get(
-        f"https://thunderdome.dev/api/battles/{battle_id}",
-        timeout=10,
-        headers=thunderdome_headers,
-    )
-
-    if not thunderdome_response.ok:
-        logging.error("Failed to fetch battle %s", battle_id)
-        return []
-
-    payload = thunderdome_response.json()
-
-    logging.debug("Fetched %d plans", len(payload["data"]["plans"]))
-
-    return payload["data"]["plans"]
-
-
-def transfer_points(
-    plans: list[dict],
-    gitlab_token: str,
-    overwrite: bool = False
-) -> None:
-    """
-    Transfer points to GitLab issues.
-
-    :param plans: Plans from the Thunderdome game.
-    :param gitlab_token: Token for the GitLab API.
-    :param overwrite: True to overwrite existing weights, False to preserve existing weights.
-    """
-    logging.info("Transferring points to GitLab...")
-
-    gitlab_headers = {
-        "PRIVATE-TOKEN": gitlab_token,
-    }
-
-    for plan in plans:
-        points = plan["points"]
-        if not points:
-            logging.warning("Skipping plan %s: No points set for plan", plan["id"])
-            continue
-
-        try:
-            _ = int(points)
-
-        except ValueError:
-            logging.error(
-                "Skipping plan %s: Points is not an integer, found '%s' instead",
-                plan["id"],
-                points,
-            )
-            continue
-
-        link = plan["link"]
-        if not link:
-            logging.warning("Skipping plan %s: No link set for plan", plan["id"])
-            continue
-
-        match = re.match(GITLAB_ISSUE_URL_REGEX, plan["link"])
-        if not match:
-            logging.error(
-                (
-                    "Skipping plan %s: Invalid URL '%s' does "
-                    "not match GitLab URL pattern '%s'"
-                ),
-                plan["id"],
-                plan["link"],
-                GITLAB_ISSUE_URL_REGEX.pattern,
-            )
-            continue
-
-        project_path = match.group("project")
-        issue_iid = match.group("issue")
-
-        # Get project ID
-        project_id = get_project_id(link, gitlab_token, GITLAB_ISSUE_URL_REGEX)
-
-        # Get issue information
-        gitlab_response = requests.get(
-            f"https://gitlab.com/api/v4/projects/{project_id}/issues/{issue_iid}",
-            timeout=10,
-            headers=gitlab_headers,
-        )
-        payload = gitlab_response.json()
-
-        if not gitlab_response.ok:
-            logging.error("Failed to fetch issue %s#%s", project_path, issue_iid)
-            continue
-
-        if not "weight" in payload:
-            logging.error(
-                "No 'weight' for issue in API response. Are you authenticated?"
-            )
-            continue
-
-        previous_weight = payload["weight"]
-
-        if previous_weight is not None and overwrite is False:
-            logging.info(
-                "Skipping %s#%s: Issue already has a weight set",
-                project_path,
-                issue_iid,
-            )
-            continue
-
-        # Set weight
-        gitlab_response = requests.put(
-            f"https://gitlab.com/api/v4/projects/{project_id}/issues/{issue_iid}",
-            timeout=10,
-            headers=gitlab_headers,
-            json={"weight": points},
-        )
-
-        if not gitlab_response.ok:
-            logging.error("Failed to set weight for %s#%s", project_path, issue_iid)
-
-        else:
-            if previous_weight is not None:
-                logging.info(
-                    "Changed weight to %s for %s#%s (was %s)",
-                    points,
-                    project_path,
-                    issue_iid,
-                    previous_weight,
-                )
-            else:
-                logging.info(
-                    "Set weight to %s for %s#%s", points, project_path, issue_iid
-                )
 
 
 def create_game(plans: list[dict], args: argparse.Namespace) -> None:
